pyproject.toml
[project]
name = "bitfa-trader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "celery>=5.6.2",
    "django>=6.0.2",
    "groq>=1.0.0",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
    "redis>=7.2.0",
    "telethon>=1.42.0",
]
